feat: add component to closed table
