+++ apps/trading/components/market-table-actions/index.ts
@@ -1,0 +1,1 @@
+export * from './market-table-actions';
